Cleaned duplicate experience definitions
Summary:
- testingconstants.js has all the experience code stored in CONSTANTS
- createNewExperiences defines the meteor methods to insert the
experiences into the DB
- Deleted Cheers (as it was basically handled by Bumped)
- Commented out 2 spooky storytimes, with followup FIXME/TODOs
